State: move ui.currentPreviewUrl to ui.preview.currentPreviewUrl
Adds a state.ui.preview tree to make it easier to add more preview-related UI
state reducers in the future.
