+++ client/state/ui/preview/actions.js
@@ -1,0 +1,20 @@
+/**
+ * Internal dependencies
+ */
+import {
+	PREVIEW_URL_CLEAR,
+	PREVIEW_URL_SET,
+} from 'state/action-types';
+
+export function setPreviewUrl( url ) {
+	return {
+		type: PREVIEW_URL_SET,
+		url,
+	};
+}
+
+export function clearPreviewUrl() {
+	return {
+		type: PREVIEW_URL_CLEAR,
+	};
+}
