+++ client/state/ui/preview/selectors.js
@@ -1,0 +1,12 @@
+/**
+ * Returns the URL if SitePreview currently has one.
+ *
+ * @param  {Object}  state Global state tree
+ * @return {?String}  The url or null
+ *
+ * @see client/components/design-preview
+ */
+export function getPreviewUrl( state ) {
+	return state.ui.preview.currentPreviewUrl;
+}
+
