+++ client/state/ui/reducer.js
@@ -10,8 +10,6 @@ import {
 	SELECTED_SITE_SET,
 	SECTION_SET,
 	PREVIEW_IS_SHOWING,
-	PREVIEW_URL_CLEAR,
-	PREVIEW_URL_SET,
 	SERIALIZE,
 	DESERIALIZE,
 } from 'state/action-types';
@@ -23,6 +21,7 @@ import reader from './reader/reducer';
 import olark from './olark/reducer';
 import actionLog from './action-log/reducer';
 import layoutFocus from './layout-focus/reducer';
+import preview from './preview/reducer';
 
 /**
  * Tracks the currently selected site ID.
@@ -90,23 +89,12 @@ export const isPreviewShowing = createReducer( false, {
 		isShowing !== undefined ? isShowing : state,
 } );
 
-export function currentPreviewUrl( state = null, action ) {
-	switch ( action.type ) {
-		case PREVIEW_URL_SET:
-			return action.url;
-		case PREVIEW_URL_CLEAR:
-			return null;
-	}
-	return state;
-}
-
 const reducer = combineReducers( {
 	section,
 	isLoading,
 	layoutFocus,
 	hasSidebar,
 	isPreviewShowing,
-	currentPreviewUrl,
 	queryArguments,
 	selectedSiteId,
 	recentlySelectedSiteIds,
@@ -114,6 +102,7 @@ const reducer = combineReducers( {
 	editor,
 	reader,
 	olark,
+	preview,
 	actionLog,
 } );
